# perf(game): denormalize leaderboard top entries