--- a/app/Platform/Commands/SyncLeaderboardTopEntries.php
+++ b/app/Platform/Commands/SyncLeaderboardTopEntries.php
@@ -34,7 +34,7 @@ class SyncLeaderboardTopEntries extends Command
         $query->chunkById(5000, function ($leaderboards) use ($bar) {
             /** @var Leaderboard $leaderboard */
             foreach ($leaderboards as $leaderboard) {
-                (new RecalculateLeaderboardTopEntryAction())->execute($leaderboard);
+                (new RecalculateLeaderboardTopEntryAction())->execute($leaderboard->id);
                 $bar->advance();
             }
         }, 'ID');