--- a/app/Platform/Listeners/RecalculateLeaderboardTopEntriesForUser.php
+++ b/app/Platform/Listeners/RecalculateLeaderboardTopEntriesForUser.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Platform\Listeners;
 
+use App\Events\UserDeleted;
 use App\Models\User;
 use App\Platform\Actions\RecalculateLeaderboardTopEntryAction;
 use App\Platform\Events\PlayerRankedStatusChanged;
@@ -19,6 +20,9 @@ class RecalculateLeaderboardTopEntriesForUser implements ShouldQueue
 
         switch ($event::class) {
             case PlayerRankedStatusChanged::class:
+                $user = $event->user;
+                break;
+            case UserDeleted::class:
                 $user = $event->user;
                 break;
         }