This PR significantly improves the load performance of several of the most popular game pages.

Most query time on game pages right now is determining the top entry for the game's various leaderboards. This PR adds one new column to `LeaderboardDef`:
* top_entry_id

With these additions, it's no longer necessary to iterate through all a leaderboard's entries to determine the top one to display on game pages.

On game ID 228 (Super Mario World), query execution time goes from ~340ms to ~110ms.
On game ID 1446 (Super Mario Bros.), query execution time goes from ~187ms to ~68ms.

**How to test this PR:**
```
sail artisan ra:sync:leaderboard-top-entries
sail artisan horizon
```
Note, **there is no cutover period for the denormalized data.** Game pages (and the web API) immediately start looking for the data - but it only takes about a minute or two to populate.

When a user is deleted or untracked, all the leaderboard top entries they're top-ranked on are automatically updated.